More license updates.
